More pagination buttons on posts page
Formerly, the top of the sandbox held a button to load the previous
page, and it loaded 100 posts, which was too many to be convenient. For
volumes that base their posts on age (instead of manual insertion, like
ordinary story-volumes do), this button is replaced with three buttons,
allowing the user to load 5, 10, or 100 previous posts.