posts#update updates Volume if most recent
When a Post is updated, it would formerly update the Volume's
updated_by_name and updated_by_id, but I don't want that to happen
unless the updated post is the most recent post (because otherwise the
Volumes page will show that the Volume was most recently updated by
someone _other_ than the person who made the most recent post, and that
can be confusing to users.) The new behaviour which this revision introduces makes the Volume update only if the Volume's most recent Post
is the updated Post.

This revision does not interfere with the updates to the Volume which
take place when a Post is created (as opposed to updated).