Notifications and time-incremental polling
Add support for Notifications.

If the polling checkbox is ticked in the user cp, then the page polls
the server for updates. Polling starts at 5 minutes and stays there, but
if any poll returns one or more posts, then the interval for the next
poll drops to half a second. Each time that the polling returns zero
posts, the interval for the next poll doubles (to a maximum of 5
minutes).