Avoid polling unless at the end of a volume
If a user hasn't tried to load the end of a volume, whether because he's
reading history or is in a paged volume and has a way to go before the
known end, it makes no sense to poll for the newest posts. The reader
should manually request the next page of posts if he wants them.